backend/requirements.txt
python-dotenv>=1.0.0
pinecone-client>=2.2.1
openai>=0.27.2
PyJWT>=2.6.0
fastapi>=0.94.1
uvicorn>=0.21.0